Move authentication stuff to its own namespace
