+++ SteamKit2/SteamKit2/Steam/SteamClient/SteamClient.cs
@@ -92,7 +92,6 @@ namespace SteamKit2
             this.AddHandler( new SteamWorkshop() );
             this.AddHandler( new SteamTrading() );
             this.AddHandler( new SteamUnifiedMessages() );
-            this.AddHandler( new SteamAuthentication() );
             this.AddHandler( new SteamScreenshots() );
             this.AddHandler( new SteamMatchmaking() );
             this.AddHandler( new SteamNetworking() );
